Update SubsetByGeneIDs
SubsetByGeneIDsServer now does the subsetting as well as loading
the gene ids from the file. It also subset the gene metadata.
The return value is now a list with subset counts and metadata as
the values.
SubsetByGeneIDsOutput is a new function that creates an anchor
point for a shinyBS alert to tell the user if any of the supplied
gene ids were not found in the data.