Create heatmap module
The heatmapOutput function create a plotOutput.
heatmapServer takes the counts, gene metadata, sample info and plots
a heatmap of the counts. axis labels are plotted if there are fewer
than 100 genes and/or 48 samples.
The Server modules also uses the state of the current transform to
decide which colour palette to use for the heatmap.
For this, the transform module has been updated to return the current
transform selected using the radio buttons provided by the transform
module.