Create CountPlot module
* countPlotInput creates select inputs for selecting the gene id,
x, fill and shape variables
* countPlotOutput creates the plotOutput and a space for warnings
about the columns of the data
* countPlotServer updates the select inputs based on the current
count and sample data. Any variables with more than 6 levels are
removed from the possible variables for shape and a warning is
generated showing which columns are unavailable for shape.
It then creates a count plot for the selected gene id and x, fill
and shape variables.